requirements.txt
Authlib==0.15.3
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
et-xmlfile==1.1.0
Flask==1.1.2
gcloud==0.18.3
gensim==3.6.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gspread==3.7.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.2
httplib2==0.19.1
httpx==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jwcrypto==0.8
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.20.2
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pathy==0.5.2
Pillow==8.2.0
preshed==3.0.5
protobuf==3.16.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.3
pyenchant==3.2.0
pyparsing==2.4.7
Pyrebase4==4.4.3
python-dateutil==2.8.1
python-jwt==3.3.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.15.0
sklearn==0.0
smart-open==3.0.0
sniffio==1.2.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
thinc==8.0.3
threadpoolctl==2.1.0
tqdm==4.60.0
typer==0.3.2
urllib3==1.26.4
wasabi==0.8.2
Werkzeug==1.0.1
xlrd==2.0.1
